requirements.txt
pandas
tensorflow
matplotlib
kafka-python
joblib
scikit-learn